+++ terratest/basic/variables.tf
@@ -33,7 +33,7 @@ variable "vswitch_ids" {
   default     = []
 }
 
-variable "instance_name" {
+variable "name" {
   description = "Name used on all instances as prefix. Like TF-ECS-Instance-1, TF-ECS-Instance-2."
   default     = "TF-ECS-Instance"
 }